requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn==22.0.0
google-generativeai==0.7.1
tenacity==8.4.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.30
blinker==1.8.2
pytest==7.4.0